server/requirements.txt
tensorflow==2.20.0
opencv-python==4.8.1.78
numpy>=1.26.0,<2.0
Flask==2.3.3
pycryptodome==3.18.0
Pillow==10.0.0
scikit-learn==1.3.2